+++ src/features/booth/components/BoothInfo.tsx
@@ -1,6 +1,8 @@
 import * as S from './BoothInfo.styles';
+import { FavoriteButton } from '@/components/favorite-button';
 import { newlineToBr } from '@/utils/newlineToBr';
 import { useBooth } from '@/hooks/useBooth';
+import { useFavorites } from '@/hooks/useFavorites';
 import { useNavigate } from 'react-router-dom';
 import { OPERATING_HOURS } from '@/constants/booth/operating-hours';
 import PubBeerIcon from '@/assets/icons/pub_beer.svg?react';
@@ -8,6 +10,7 @@ import TimeIcon from '@/assets/icons/time_pub.svg?react';
 
 export default function BoothInfo({ id }: { id: number }) {
   const { booth, error } = useBooth(id);
+  const { handleToggleFavorite, isFavorited } = useFavorites();
   const navigate = useNavigate();
 
   if (error || !booth) {
@@ -21,10 +24,20 @@ export default function BoothInfo({ id }: { id: number }) {
     });
     return null;
   }
+
+  const isBoothFavorited = isFavorited(booth.id);
+
   return (
     <S.Container>
       <S.ImageBtnFrame>
         <S.Image src={booth.profileImage} />
+        <FavoriteButton
+          id={booth.id}
+          isFavorited={isBoothFavorited}
+          onClick={handleToggleFavorite}
+          variant="large"
+          position="relative"
+        />
       </S.ImageBtnFrame>
       <S.TextSection>
         <S.TextFrame>
